requirements/app.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements/app.txt requirements/app.in
#
alembic==1.7.4
    # via flask-migrate
altair==4.1.0
    # via
    #   -r requirements/app.in
    #   timely-beliefs
arrow==1.2.1
    # via rq-dashboard
async-generator==1.10
    # via
    #   trio
    #   trio-websocket
attrs==21.2.0
    # via
    #   jsonschema
    #   outcome
    #   trio
babel==2.9.1
    # via py-moneyed
bcrypt==3.2.0
    # via -r requirements/app.in
beautifulsoup4==4.10.0
    # via siphon
blinker==1.4
    # via
    #   flask-mail
    #   flask-principal
    #   flask-security-too
    #   sentry-sdk
bokeh==1.0.4
    # via
    #   -r requirements/app.in
    #   pandas-bokeh
certifi==2021.10.8
    # via
    #   requests
    #   sentry-sdk
    #   urllib3
cffi==1.15.0
    # via
    #   bcrypt
    #   cryptography
cftime==1.5.1.1
    # via netcdf4
charset-normalizer==2.0.7
    # via requests
click==8.0.3
    # via
    #   -r requirements/app.in
    #   flask
    #   rq
colour==0.1.5
    # via -r requirements/app.in
cryptography==35.0.0
    # via
    #   pyopenssl
    #   urllib3
cycler==0.11.0
    # via matplotlib
dnspython==2.1.0
    # via email-validator
email-validator==1.1.3
    # via
    #   -r requirements/app.in
    #   flask-security-too
entrypoints==0.3
    # via altair
filelock==3.3.2
    # via tldextract
flask==2.0.2
    # via
    #   -r requirements/app.in
    #   flask-classful
    #   flask-cors
    #   flask-json
    #   flask-login
    #   flask-mail
    #   flask-marshmallow
    #   flask-migrate
    #   flask-principal
    #   flask-security-too
    #   flask-sqlalchemy
    #   flask-sslify
    #   flask-wtf
    #   rq-dashboard
    #   sentry-sdk
flask-classful==0.14.2
    # via -r requirements/app.in
flask-cors==3.0.10
    # via -r requirements/app.in
flask-json==0.3.4
    # via -r requirements/app.in
flask-login==0.5.0
    # via
    #   -r requirements/app.in
    #   flask-security-too
flask-mail==0.9.1
    # via -r requirements/app.in
flask-marshmallow==0.14.0
    # via -r requirements/app.in
flask-migrate==3.1.0
    # via -r requirements/app.in
flask-principal==0.4.0
    # via flask-security-too
flask-security-too==4.1.2
    # via -r requirements/app.in
flask-sqlalchemy==2.5.1
    # via
    #   -r requirements/app.in
    #   flask-migrate
flask-sslify==0.1.5
    # via -r requirements/app.in
flask-wtf==0.15.1
    # via
    #   -r requirements/app.in
    #   flask-security-too
greenlet==1.1.2
    # via sqlalchemy
h11==0.12.0
    # via wsproto
h5py==3.5.0
    # via pvlib
humanize==3.12.0
    # via -r requirements/app.in
idna==3.3
    # via
    #   email-validator
    #   requests
    #   tldextract
    #   trio
    #   urllib3
importlib-metadata==4.8.1
    # via
    #   -r requirements/app.in
    #   timely-beliefs
inflect==5.3.0
    # via -r requirements/app.in
inflection==0.5.1
    # via -r requirements/app.in
iso8601==0.1.16
    # via -r requirements/app.in
isodate==0.6.0
    # via
    #   -r requirements/app.in
    #   timely-beliefs
itsdangerous==2.0.1
    # via
    #   flask
    #   flask-security-too
    #   flask-wtf
jinja2==3.0.2
    # via
    #   altair
    #   bokeh
    #   flask
joblib==1.1.0
    # via scikit-learn
jsonschema==4.1.2
    # via altair
kiwisolver==1.3.2
    # via matplotlib
mako==1.1.5
    # via alembic
markupsafe==2.0.1
    # via
    #   jinja2
    #   mako
    #   wtforms
marshmallow==3.14.0
    # via
    #   -r requirements/app.in
    #   flask-marshmallow
    #   marshmallow-polyfield
    #   marshmallow-sqlalchemy
    #   webargs
marshmallow-polyfield==5.10
    # via -r requirements/app.in
marshmallow-sqlalchemy==0.26.1
    # via -r requirements/app.in
matplotlib==3.4.3
    # via timetomodel
netcdf4==1.5.8
    # via -r requirements/app.in
numexpr==2.7.3
    # via tables
numpy==1.21.3
    # via
    #   -r requirements/app.in
    #   altair
    #   bokeh
    #   cftime
    #   h5py
    #   matplotlib
    #   netcdf4
    #   numexpr
    #   pandas
    #   patsy
    #   properscoring
    #   pvlib
    #   scikit-learn
    #   scipy
    #   siphon
    #   statsmodels
    #   tables
    #   timely-beliefs
    #   timetomodel
openturns==1.17
    # via timely-beliefs
outcome==1.1.0
    # via trio
packaging==21.2
    # via
    #   bokeh
    #   pint
pandas==1.2.5
    # via
    #   -r requirements/app.in
    #   altair
    #   pandas-bokeh
    #   pvlib
    #   siphon
    #   statsmodels
    #   timely-beliefs
    #   timetomodel
pandas-bokeh==0.4.3
    # via -r requirements/app.in
passlib==1.7.4
    # via flask-security-too
patsy==0.5.2
    # via statsmodels
pillow==8.4.0
    # via
    #   bokeh
    #   matplotlib
pint==0.18
    # via -r requirements/app.in
ply==3.11
    # via pyomo
properscoring==0.1
    # via timely-beliefs
protobuf==3.19.1
    # via siphon
pscript==0.7.6
    # via -r requirements/app.in
psycopg2-binary==2.9.1
    # via
    #   -r requirements/app.in
    #   timely-beliefs
pvlib==0.9.0
    # via -r requirements/app.in
py-moneyed==2.0
    # via -r requirements/app.in
pycparser==2.20
    # via cffi
pyomo==6.1.2
    # via -r requirements/app.in
pyopenssl==21.0.0
    # via urllib3
pyparsing==2.4.7
    # via
    #   matplotlib
    #   packaging
pyrsistent==0.18.0
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   arrow
    #   bokeh
    #   matplotlib
    #   pandas
    #   timetomodel
python-dotenv==0.19.1
    # via -r requirements/app.in
pytz==2021.3
    # via
    #   -r requirements/app.in
    #   babel
    #   pandas
    #   pvlib
    #   timely-beliefs
    #   timetomodel
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0
    # via bokeh
redis==3.5.3
    # via
    #   rq
    #   rq-dashboard
requests==2.26.0
    # via
    #   pvlib
    #   requests-file
    #   siphon
    #   tldextract
requests-file==1.5.1
    # via tldextract
rq==1.10.0
    # via
    #   -r requirements/app.in
    #   rq-dashboard
rq-dashboard==0.6.1
    # via -r requirements/app.in
scikit-learn==1.0.1
    # via sklearn
scipy==1.7.1
    # via
    #   properscoring
    #   pvlib
    #   scikit-learn
    #   statsmodels
    #   timely-beliefs
    #   timetomodel
selenium==4.0.0
    # via timely-beliefs
sentry-sdk[flask]==1.4.3
    # via -r requirements/app.in
siphon==0.9
    # via -r requirements/app.in
six==1.16.0
    # via
    #   bcrypt
    #   bokeh
    #   flask-cors
    #   flask-marshmallow
    #   isodate
    #   patsy
    #   pyopenssl
    #   python-dateutil
    #   requests-file
sklearn==0.0
    # via timetomodel
sniffio==1.2.0
    # via trio
sortedcontainers==2.4.0
    # via trio
soupsieve==2.2.1
    # via beautifulsoup4
sqlalchemy==1.4.26
    # via
    #   -r requirements/app.in
    #   alembic
    #   flask-sqlalchemy
    #   marshmallow-sqlalchemy
    #   timely-beliefs
    #   timetomodel
statsmodels==0.13.0
    # via timetomodel
tables==3.6.1
    # via -r requirements/app.in
threadpoolctl==3.0.0
    # via scikit-learn
timely-beliefs==1.9.2
    # via -r requirements/app.in
timetomodel==0.7.1
    # via -r requirements/app.in
tldextract==3.1.2
    # via -r requirements/app.in
toolz==0.11.1
    # via altair
tornado==6.1
    # via bokeh
trio==0.19.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.9.2
    # via selenium
typing-extensions==4.0.1
    # via py-moneyed
tzdata==2021.5
    # via pytz-deprecation-shim
tzlocal==4.1
    # via -r requirements/app.in
urllib3[secure]==1.26.7
    # via
    #   requests
    #   selenium
    #   sentry-sdk
webargs==8.0.1
    # via -r requirements/app.in
werkzeug==2.0.2
    # via flask
wsproto==1.0.0
    # via trio-websocket
wtforms==2.3.3
    # via flask-wtf
xlrd==2.0.1
    # via -r requirements/app.in
zipp==3.6.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
